+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Unreleased
 
 ## Features
-* routing -  a template argument was added to the `@routing.route` decorator.
+* routing - a template argument was added to the `@routing.route` decorator.
   This argument determines which templates a route can be added to.
   https://github.com/anvilistas/anvil-extras/issues/293
+* routing - a tempalate can take multiple paths `@routing.template(path=["admin", "user"])`
+  https://github.com/anvilistas/anvil-extras/pull/298
+* routing - `@routing.redirect()` decorator added
+  https://github.com/anvilistas/anvil-extras/pull/298
+
 
 
 # v2.0.1 16-Mar-2022